moved some functions to their proper entity type
