+++ src/main/java/net/blancworks/figura/lua/api/actionWheel/ActionWheelAPI.java
@@ -48,6 +48,7 @@ public class ActionWheelAPI {
                     return NIL;
                 }
             });
+
             set("setRightSize", new OneArgFunction() {
                 @Override
                 public LuaValue call(LuaValue arg) {
@@ -56,25 +57,29 @@ public class ActionWheelAPI {
                     return NIL;
                 }
             });
+
             set("getLeftSize", new ZeroArgFunction() {
                 @Override
                 public LuaValue call() {
                     return LuaValue.valueOf(script.actionWheelLeftSize);
                 }
             });
+
             set("getRightSize", new ZeroArgFunction() {
                 @Override
                 public LuaValue call() {
                     return LuaValue.valueOf(script.actionWheelRightSize);
                 }
             });
+
             set("getSelectedSlot", new ZeroArgFunction() {
                 @Override
                 public LuaValue call() {
                     int selected = ActionWheel.selectedSlot;
-                    return selected == -1 ? NIL : LuaValue.valueOf(selected+1);
+                    return selected == -1 ? NIL : LuaValue.valueOf(selected + 1);
                 }
             });
+
         }});
     }
