+++ src/main/java/net/blancworks/figura/lua/api/world/entity/LivingEntityAPI.java
@@ -1,12 +1,9 @@
 package net.blancworks.figura.lua.api.world.entity;
 
-import net.blancworks.figura.lua.api.math.LuaVector;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.effect.StatusEffect;
 import net.minecraft.entity.effect.StatusEffectInstance;
 import net.minecraft.util.Identifier;
-import net.minecraft.util.hit.BlockHitResult;
-import net.minecraft.util.hit.HitResult;
 import net.minecraft.util.registry.Registry;
 import org.luaj.vm2.LuaNumber;
 import org.luaj.vm2.LuaString;
@@ -126,32 +123,6 @@ public class LivingEntityAPI {
                 @Override
                 public LuaValue call() {
                     return LuaNumber.valueOf(targetEntity.get().getStingerCount());
-                }
-            });
-
-            superTable.set("getFireTicks", new ZeroArgFunction() {
-                @Override
-                public LuaValue call() {
-                    return LuaNumber.valueOf(targetEntity.get().getFireTicks());
-                }
-            });
-
-            superTable.set("getFrozenTicks", new ZeroArgFunction() {
-                @Override
-                public LuaValue call() {
-                    return LuaNumber.valueOf(targetEntity.get().getFrozenTicks());
-                }
-            });
-
-            superTable.set("getTargetedBlockPos", new OneArgFunction() {
-                @Override
-                public LuaValue call(LuaValue arg) {
-                    HitResult result = targetEntity.get().raycast(20.0D, 0.0F, arg.checkboolean());
-                    if (result.getType() == HitResult.Type.BLOCK) {
-                        return LuaVector.of(((BlockHitResult)result).getBlockPos());
-                    } else {
-                        return NIL;
-                    }
                 }
             });
