+++ src/main/java/net/blancworks/figura/lua/api/world/entity/PlayerEntityAPI.java
@@ -6,6 +6,7 @@ import net.blancworks.figura.lua.api.item.ItemStackAPI;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.network.ClientPlayerEntity;
 import net.minecraft.entity.Entity;
+import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.Identifier;
@@ -99,10 +100,20 @@ public class PlayerEntityAPI {
                 public LuaValue call() {
                     if (targetEntity.get() instanceof ClientPlayerEntity) {
                         Entity lookingAt = MinecraftClient.getInstance().targetedEntity;
+
                         if (lookingAt != null && !lookingAt.isInvisibleTo(targetEntity.get())) {
-                            return new EntityAPI.EntityLuaAPITable(() -> lookingAt).getTable();
+                            if (lookingAt instanceof PlayerEntity) {
+                                return new PlayerEntityLuaAPITable(() -> (PlayerEntity) lookingAt).getTable();
+                            }
+                            else if (lookingAt instanceof LivingEntity) {
+                                return new LivingEntityAPI.LivingEntityAPITable<>(() -> (LivingEntity) lookingAt).getTable();
+                            }
+                            else {
+                                return new EntityAPI.EntityLuaAPITable<>(() -> lookingAt).getTable();
+                            }
                         }
                     }
+
                     return NIL;
                 }
             });
